requirements.txt
pyyaml
requests
minio
pymbtiles
python-dotenv
schedule
python-slugify
